[Major] Fixed the memory bottleneck of data processing

+++ quantize/block_ap.py
@@ -19,13 +19,15 @@ from torch.utils.data import DataLoader
 import shutil
 import os
 
-def update_dataloader(layer, dataset, dev, attention_mask, position_ids, offload=False):
+def update_dataset(layer, dataset, dev, attention_mask, position_ids):
     with torch.no_grad():
         with torch.cuda.amp.autocast():
-            for index1, inps in enumerate(dataset):
-                inps = inps.to(dev).unsqueeze(0)
+            for index, inps in enumerate(dataset):
+                inps = inps.to(dev)
+                if len(inps.shape)==2:
+                    inps = inps.unsqueeze(0)
                 new_data = layer(inps, attention_mask=attention_mask,position_ids=position_ids)[0].to('cpu')
-                dataset.update_data(index1,new_data)
+                dataset.update_data(index,new_data)
 
                     
 def block_ap(
@@ -43,104 +45,80 @@ def block_ap(
     use_cache = model.config.use_cache
     model.config.use_cache = False
     
-    # move embedding layer and first layer to target device
-    # only suppress llama models now
+    # step 1: move embedding layer and first layer to target device, only suppress llama models now
     layers = model.model.layers
     model.model.embed_tokens = model.model.embed_tokens.to(dev)
     model.model.norm = model.model.norm.to(dev)
     if hasattr(model.model, 'rotary_emb'):
         # for llama-3.1
         model.model.rotary_emb = model.model.rotary_emb.to(dev)
-    is_llama = True     
-    
-    
     layers[0] = layers[0].to(dev)
     dtype = torch.float16
 
-    # path to save intermediate data
+    # step 2: init dataset
     flag = time.time()
-    fp_train_cache_path = f'{args.cache_dir}/{flag}/block_training_fp_train'
-    fp_val_cache_path = f'{args.cache_dir}/{flag}/block_training_fp_val'
-    quant_train_cache_path = f'{args.cache_dir}/{flag}/block_training_quant_train'
-    quant_val_cache_path = f'{args.cache_dir}/{flag}/block_training_quant_val'
-    for path in [fp_train_cache_path,fp_val_cache_path,quant_train_cache_path,quant_val_cache_path]:
-        if os.path.exists(path):
-            shutil.rmtree(path)
-
+    if args.off_load_to_disk:
+        fp_train_cache_path = f'{args.cache_dir}/{flag}/block_training_fp_train'
+        fp_val_cache_path = f'{args.cache_dir}/{flag}/block_training_fp_val'
+        quant_train_cache_path = f'{args.cache_dir}/{flag}/block_training_quant_train'
+        quant_val_cache_path = f'{args.cache_dir}/{flag}/block_training_quant_val'
+        for path in [fp_train_cache_path,fp_val_cache_path,quant_train_cache_path,quant_val_cache_path]:
+            if os.path.exists(path):
+                shutil.rmtree(path)
+    else:
+        fp_train_cache_path = None
+        fp_val_cache_path = None
+        quant_train_cache_path = None
+        quant_val_cache_path = None
     fp_train_inps = BlockTrainDataset(args.train_size, args.training_seqlen, 
-                                model.config.hidden_size, dtype, cache_path=fp_train_cache_path,off_load_to_disk=args.off_load_to_disk)
+                                model.config.hidden_size, args.batch_size, dtype, cache_path=fp_train_cache_path,off_load_to_disk=args.off_load_to_disk)
     fp_val_inps = BlockTrainDataset(args.val_size, args.training_seqlen, 
-                                model.config.hidden_size, dtype, cache_path=fp_val_cache_path,off_load_to_disk=args.off_load_to_disk)
-    fp_train_inps_loader = DataLoader(fp_train_inps, batch_size=args.batch_size, shuffle=False, num_workers=args.num_workers,prefetch_factor=args.prefetch_factor)
-    fp_val_inps_loader = DataLoader(fp_val_inps, batch_size=args.batch_size, shuffle=False, num_workers=args.num_workers,prefetch_factor=args.prefetch_factor)
+                                model.config.hidden_size, args.batch_size, dtype, cache_path=fp_val_cache_path,off_load_to_disk=args.off_load_to_disk)
     
-    # catch the first layer input
-    cache = {"i": 0}
+    # step 3: catch the input of thefirst layer 
     class Catcher(nn.Module):
-        def __init__(self, module):
+        def __init__(self, module, dataset):
             super().__init__()
             self.module = module
-            self.is_llama = False
+            self.dataset = dataset
+            self.index = 0
+            self.attention_mask = None
+            self.position_ids = None
 
         def forward(self, inp, **kwargs):
-            if cache["i"]<=args.train_size-1:
-                fp_train_inps.update_data(cache["i"], inp.squeeze(0).to('cpu'))
-            else:
-                fp_val_inps.update_data(cache["i"]-args.train_size, inp.squeeze(0).to('cpu'))
-            cache["i"] += 1
-            cache["attention_mask"] = kwargs["attention_mask"]
-            if self.is_llama:
-                cache["position_ids"] = kwargs["position_ids"]
+            self.dataset.update_data(self.index, inp.squeeze(0).to('cpu'))
+            self.index += 1
+            if self.attention_mask is None:
+                self.attention_mask = kwargs["attention_mask"]
+            if self.position_ids is None:
+                self.position_ids = kwargs["position_ids"]
             raise ValueError
     
-
-
-    layers[0] = Catcher(layers[0])
-    layers[0].is_llama = is_llama
-
-
+    # step 3.1: catch the input of training set
+    layers[0] = Catcher(layers[0],fp_train_inps)
+    iters = len(trainloader)//args.batch_size
     with torch.no_grad():
-        for batch in (trainloader+valloader):
-            if cache["i"] >= args.train_size + args.val_size:
-                break
+        for i in range(iters):
+            data = torch.cat([trainloader[j][0] for j in range(i*args.batch_size,(i+1)*args.batch_size)],dim=0)
             try:
-                model(batch[0].to(dev))
+                model(data.to(dev))
             except ValueError:
                 pass
-    
-    # move embedding layer and first layer to cpu
     layers[0] = layers[0].module
-    layers[0] = layers[0].cpu()
-    model.model.embed_tokens = model.model.embed_tokens.cpu()
-    model.model.norm = model.model.norm.cpu()
-    if hasattr(model.model, 'rotary_emb'):
-        # for llama-3.1
-        model.model.rotary_emb = model.model.rotary_emb.cpu()
-    torch.cuda.empty_cache()
 
-    if args.off_load_to_disk:
-        # copy quant input from fp input, they are same in first layer
-        shutil.copytree(fp_train_cache_path, quant_train_cache_path)
-        shutil.copytree(fp_val_cache_path, quant_val_cache_path)
-        quant_train_inps = BlockTrainDataset(args.train_size, args.training_seqlen, 
-                                    model.config.hidden_size, dtype, cache_path=quant_train_cache_path,off_load_to_disk=args.off_load_to_disk)
-        quant_val_inps = BlockTrainDataset(args.val_size, args.training_seqlen, 
-                                    model.config.hidden_size, dtype, cache_path=quant_val_cache_path,off_load_to_disk=args.off_load_to_disk)
-    else:
-        quant_train_inps = BlockTrainDataset(args.train_size, args.training_seqlen, 
-                                    model.config.hidden_size, dtype, cache_path=quant_train_cache_path,off_load_to_disk=args.off_load_to_disk)
-        quant_val_inps = BlockTrainDataset(args.val_size, args.training_seqlen, 
-                                    model.config.hidden_size, dtype, cache_path=quant_val_cache_path,off_load_to_disk=args.off_load_to_disk)
-        for index,data in enumerate(fp_train_inps):
-            quant_train_inps.update_data(index, data)
-        for index,data in enumerate(fp_val_inps):
-            quant_val_inps.update_data(index, data)
-            
-    quant_train_inps_loader = DataLoader(quant_train_inps, batch_size=args.batch_size, shuffle=False, num_workers=args.num_workers,prefetch_factor=args.prefetch_factor)
-    quant_val_inps_loader = DataLoader(quant_val_inps, batch_size=args.batch_size, shuffle=False, num_workers=args.num_workers,prefetch_factor=args.prefetch_factor)
-
-    
-    attention_mask = cache["attention_mask"]
+    # step 3.2: catch the input of validation set
+    layers[0] = Catcher(layers[0],fp_val_inps)
+    iters = len(valloader)//args.batch_size
+    with torch.no_grad():
+        for i in range(iters):
+            data = torch.cat([valloader[j][0] for j in range(i*args.batch_size,(i+1)*args.batch_size)],dim=0)
+            try:
+                model(data.to(dev))
+            except ValueError:
+                pass
+    attention_mask = layers[0].attention_mask
+    position_ids = layers[0].position_ids
+    layers[0] = layers[0].module
     if attention_mask is not None:
         attention_mask_batch = attention_mask.repeat(args.batch_size,1,1,1).float()
     else:
@@ -149,47 +127,66 @@ def block_ap(
             " Seems that model's attention works without a mask."
         )
         attention_mask_batch = None
+    
+    # step 4: move embedding layer and first layer to cpu
+    layers[0] = layers[0].cpu()
+    model.model.embed_tokens = model.model.embed_tokens.cpu()
+    model.model.norm = model.model.norm.cpu()
+    if hasattr(model.model, 'rotary_emb'):
+        # for llama-3.1
+        model.model.rotary_emb = model.model.rotary_emb.cpu()
+    torch.cuda.empty_cache()
 
-    loss_func = torch.nn.MSELoss()
-    if is_llama:
-        position_ids = cache["position_ids"]
+    # step 5: copy fp input as the quant input, they are same at the first layer
+    if args.off_load_to_disk:
+        # copy quant input from fp input, they are same in first layer
+        shutil.copytree(fp_train_cache_path, quant_train_cache_path)
+        shutil.copytree(fp_val_cache_path, quant_val_cache_path)
+        quant_train_inps = BlockTrainDataset(args.train_size, args.training_seqlen, 
+                                    model.config.hidden_size, args.batch_size, dtype, cache_path=quant_train_cache_path,off_load_to_disk=args.off_load_to_disk)
+        quant_val_inps = BlockTrainDataset(args.val_size, args.training_seqlen, 
+                                    model.config.hidden_size, args.batch_size, dtype, cache_path=quant_val_cache_path,off_load_to_disk=args.off_load_to_disk)
     else:
-        position_ids = None
-    
-    
+        quant_train_inps = BlockTrainDataset(args.train_size, args.training_seqlen, 
+                                    model.config.hidden_size, args.batch_size, dtype, cache_path=quant_train_cache_path,off_load_to_disk=args.off_load_to_disk)
+        quant_val_inps = BlockTrainDataset(args.val_size, args.training_seqlen, 
+                                    model.config.hidden_size, args.batch_size, dtype, cache_path=quant_val_cache_path,off_load_to_disk=args.off_load_to_disk)
+        for index,data in enumerate(fp_train_inps):
+            quant_train_inps.update_data(index, data)
+        for index,data in enumerate(fp_val_inps):
+            quant_val_inps.update_data(index, data)
 
+    # step 6: start training    
+    loss_func = torch.nn.MSELoss()
     for block_index in range(len(layers)):
         logger.info(f"=== Start quantize blocks {block_index}===")
+        # step 6.1: replace torch.nn.Linear with QuantLinear for QAT
         layer = layers[block_index].to(dev)
         qlayer = copy.deepcopy(layer)
-        # replace torch.nn.Linear with QuantLinear for QAT
         for name, module in qlayer.named_modules():
             if isinstance(module,torch.nn.Linear):
                 quantlinear = int_linear_fake.QuantLinear(module, args.wbits, args.group_size)
                 set_op_by_name(qlayer, name, quantlinear)  
                 del module  
-        
-        
-        # obtain output of full-precision model
         qlayer.to(dev)
-        # deactivate quantization for obtaining ground truth
-        set_quant_state(qlayer,weight_quant=False)
-        if args.epochs > 0:
-            update_dataloader(qlayer,fp_train_inps,dev,attention_mask,position_ids)
-            update_dataloader(qlayer,fp_val_inps,dev,attention_mask,position_ids)
-
-        # activate quantization
-        set_quant_state(qlayer,weight_quant=True)  
         
-        total_training_iteration = args.epochs * args.train_size / args.batch_size
+        
+        # step 6.2: obtain output of full-precision model for MSE
+        set_quant_state(qlayer,weight_quant=False) # deactivate quantization for obtaining ground truth
+        if args.epochs > 0:
+            update_dataset(qlayer,fp_train_inps,dev,attention_mask,position_ids)
+            update_dataset(qlayer,fp_val_inps,dev,attention_mask,position_ids)
+        set_quant_state(qlayer,weight_quant=True)  # activate quantization
+        
         
         if args.epochs > 0:
             with torch.no_grad():
                 qlayer.float()      # fp32 is required for AMP training
-            # create optimizer
+            # step 6.3: create optimizer and learning rate schedule
             param = []
             assert args.quant_lr > 0 or args.weight_lr > 0
             param_group_index = 0
+            total_training_iteration = args.epochs * args.train_size / args.batch_size 
             if args.quant_lr > 0:
                 set_quant_parameters(qlayer,True)
                 param.append({"params":quant_parameters(qlayer),"lr":args.quant_lr})
@@ -210,7 +207,6 @@ def block_ap(
             else:
                 set_weight_parameters(qlayer,False)
             optimizer = torch.optim.AdamW(param, weight_decay=args.wd)
-
             loss_scaler = utils.NativeScalerWithGradNormCount()
             trainable_number = trainable_parameters_num(qlayer)
             print(f"trainable parameter number: {trainable_number/1e6}M")
@@ -218,10 +214,11 @@ def block_ap(
             best_val_loss = 1e6
             early_stop_flag = 0
             for epoch in range(args.epochs):
+                # step: 6.4 training
                 loss_list = []
                 norm_list = []
                 start_time = time.time()
-                for index, (quant_inps, fp_inps) in enumerate(zip(quant_train_inps_loader, fp_train_inps_loader)):    
+                for index, (quant_inps, fp_inps) in enumerate(zip(quant_train_inps, fp_train_inps)):    
                     # obtain output of quantization model
                     with torch.cuda.amp.autocast():
                         input = quant_inps.to(dev)
@@ -246,8 +243,9 @@ def block_ap(
                         weight_scheduler.step()
                         optimizer.param_groups[weight_index]['lr'] = weight_scheduler.get_lr()[0]
 
+                # step 6.5: calculate validation loss
                 val_loss_list = []
-                for index, (quant_inps,fp_inps) in enumerate(zip(quant_val_inps_loader, fp_val_inps_loader)):  
+                for index, (quant_inps,fp_inps) in enumerate(zip(quant_val_inps, fp_val_inps)):  
                     # obtain output of quantization model
                     with torch.no_grad():
                         with torch.cuda.amp.autocast():
@@ -271,20 +269,18 @@ def block_ap(
             optimizer.zero_grad()
             del optimizer
 
-        # real smooth and quantization
+        # step 6.6: directly replace the weight with fake quantization
         qlayer.half()
-        # directly replace the weight with fake quantization
         quant_inplace(qlayer)
         set_quant_state(qlayer,weight_quant=False)  # weight has been quantized inplace
+
+        # step 6.7: update inputs of quantization model
         if args.epochs>0:
-            # update inputs of quantization model
-            update_dataloader(qlayer,quant_train_inps,dev,attention_mask,position_ids)
-            update_dataloader(qlayer,quant_val_inps,dev,attention_mask,position_ids)
-        
-        # move to cpu
+            update_dataset(qlayer,quant_train_inps,dev,attention_mask,position_ids)
+            update_dataset(qlayer,quant_val_inps,dev,attention_mask,position_ids)
         layers[block_index] = qlayer.to("cpu")
 
-        # pack quantized weights, note that this process is slow on poor CPU or busy CPU
+        # step 7: pack quantized weights into low-bits format, note that this process is slow on poor CPU or busy CPU
         if args.real_quant:
             named_linears = get_named_linears(qlayer, int_linear_fake.QuantLinear)
             for name, module in named_linears.items():
@@ -303,9 +299,10 @@ def block_ap(
         torch.cuda.empty_cache()
 
     # delete cached dataset
-    for path in [fp_train_cache_path,fp_val_cache_path,quant_train_cache_path,quant_val_cache_path]:
-        if os.path.exists(path):
-            shutil.rmtree(path)
+    if args.off_load_to_disk:
+        for path in [fp_train_cache_path,fp_val_cache_path,quant_train_cache_path,quant_val_cache_path]:
+            if os.path.exists(path):
+                shutil.rmtree(path)
 
     torch.cuda.empty_cache()
     gc.collect()
